pyproject.toml
[project]
name = "reid-soft-attributes"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.1",
    "pyyaml>=6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchreid>=0.2.5",
    "torchvision>=0.25.0",
]

